+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- New method `parameter_texts` for Report to display selected parameters in an array of human-readable strings.
-- New column attribute `pinned`. Can be 'left' or 'right' or nil.
 ### Changed
+### Fixed
+
+## [0.1.6] - 2019-01-08
+### Added
+- New method `parameter_texts` for Report to display selected parameters in an array of human-readable strings.
+- New Column attribute `pinned`. Can be 'left' or 'right' or nil.
 ### Fixed
 - Raise an exception if the SQL used for a QueryParameterDefinition `list_def` is not a SELECT.